+++ src/Sentry/Protocol/Envelopes/Envelope.cs
@@ -451,6 +451,18 @@ public sealed class Envelope : ISerializable, IDisposable
     internal static Envelope FromAttachment(SentryId eventId, SentryAttachment attachment, IDiagnosticLogger? logger = null) =>
         new(eventId, CreateHeader(eventId), [EnvelopeItem.FromAttachment(attachment)]);
 
+    internal static Envelope FromLog(StructuredLog log)
+    {
+        var header = DefaultHeader;
+
+        var items = new[]
+        {
+            EnvelopeItem.FromLog(log),
+        };
+
+        return new Envelope(header, items);
+    }
+
     private static async Task<IReadOnlyDictionary<string, object?>> DeserializeHeaderAsync(
         Stream stream,
         CancellationToken cancellationToken = default)